+++ tests/testthat/test-funguild_assign.R
@@ -22,10 +22,32 @@ test_that("funguild_assign accepts a character vector", {
   )
 })
 
+dbfile <- tempfile(fileext = "rds")
+
 test_that("get_funguild_db works", {
   testthat::skip_if_offline(host = "www.stbates.org")
-  assignment <- funguild_assign(sample_fungi)
-  expect_is(assignment, "tbl_df")
+  db <- get_funguild_db()
+  expect_is(db, "tbl_df")
+  expect_gt(nrow(db), 100)
+  saveRDS(db, dbfile)
+})
+
+test_that("direct assignment and stored db give same results", {
+  testthat::skip_if_offline(host = "www.stbates.org")
+  assignment_online <- funguild_assign(sample_fungi)
+  testthat::skip_if_not(file.exists(dbfile))
+  db <- readRDS(dbfile)
+  assignment_local <- funguild_assign(sample_fungi, db)
+  expect_identical(assignment_online, assignment_local)
+})
+
+test_that("direct query and stored db give same results", {
+  testthat::skip_if_offline(host = "www.mycoportal.org")
+  result_online <- funguild_query("Ectomycorrhizal*", "guild")
+  testthat::skip_if_not(file.exists(dbfile))
+  db <- readRDS(dbfile)
+  result_local <- funguild_query("Ectomycorrhizal*", "guild", db = db)
+  expect_identical(result_local, result_online)
 })
 
 sample2 <- sample_fungi